Fixed all the calls for the toolchain
