+++ DependencyFetcher/DependencyFetcherOptions.cs
@@ -9,9 +9,9 @@ namespace DependencyFetcher
 {
     public class DependencyFetcherOptions
     {
-        [Option('p', "root", Required = true, HelpText = "Path to the root to install dependencies to")]
+        [Option("root", Required = true, HelpText = "Path to the root to install dependencies to")]
         public string RootPath { get; set; } = "";
-        [Option('d', "deps", Required = true, HelpText = "Path to the dependency file to use")]
+        [Option("deps", Required = true, HelpText = "Path to the dependency file to use")]
         public string DependencyPath { get; set; } = "";
     }
 }
