+++ Toolchain/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "Toolchain": {
       "commandName": "Project",
-      "commandLineArgs": "--benchmark Benchmarks/depot.json\r\n--samples 2\r\n--downward Dependencies/fast-downward/fast-downward.py\r\n--seed 1",
+      "commandLineArgs": "--root Dependencies\r\n--deps Dependencies/dependencies.json\r\n\r\n--benchmark Benchmarks/depot.json\r\n--outdir PlanSamples\r\n--datadir Dependencies/downward-benchmarks-master\r\n--samples 2\r\n--fastdownward Dependencies/fast-downward/fast-downward.py\r\n--seed 1",
       "workingDirectory": "../"
     }
   }
