+++ Tools/BaseCLI.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using CommandLine;
+using CommandLine.Text;
 using Tools.Benchmarks;
 
 namespace Tools
@@ -12,18 +13,11 @@ namespace Tools
     {
         public static void HandleParseError(IEnumerable<Error> errs)
         {
+            var sentenceBuilder = SentenceBuilder.Create();
             foreach (var error in errs)
-                ConsoleHelper.WriteLineColor($"{error}", ConsoleColor.Red);
-        }
-
-        public static Benchmark ParseBenchmarkFile(string path)
-        {
-            ConsoleHelper.WriteLineColor("Parsing benchmark file...", ConsoleColor.DarkGray);
-            if (!File.Exists(path))
-                throw new FileNotFoundException("The given benchmark file was not found!");
-            var benchmarkFile = new Benchmark(path);
-            ConsoleHelper.WriteLineColor("Done!", ConsoleColor.Green);
-            return benchmarkFile;
+            {
+                ConsoleHelper.WriteLineColor(sentenceBuilder.FormatError(error), ConsoleColor.Red);
+            }
         }
     }
 }
